pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["SpatialTransform"]

[project]
name = "spatial-transform"
version = "1.3.3"
authors = [ { name="Wasserwecken", email="author@example.com" }, ]
description = "Transform hierarchies in 3D space."
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
  'PyGLM==2.8.2',
]

[project.urls]
"Homepage" = "https://github.com/Wasserwecken/spatial-transform"
"Bug Tracker" = "https://github.com/Wasserwecken/spatial-transform/issues"
